Remove StartupBase
Remove the separate StartupBase class and just use a single Startup class.
